requirements.txt
quart==0.18.3
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.2.3
